+++ app/public/admin/users/deleteuser.php
@@ -1,6 +1,13 @@
 <?php
 
 use Project\Services\UserService as UserService;
+use Project\Helper\Authorization;
+
+if (!Authorization::isAdmin())
+{
+    header('Location: /404/404');
+    die();
+}
 $service = new UserService();
 $data = $service->deleteUserById();
 $message = '';
